+++ io/web/server.go
@@ -3,19 +3,21 @@ package web
 import (
 	"fmt"
 	"net/http"
+	"strings"
 	"time"
 
 	rice "github.com/GeertJohan/go.rice"
+	vlib "github.com/gnames/gnlib/domain/entity/verifier"
 	"github.com/gnames/gnverify"
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 )
 
-const withLogs = false
+const withLogs = true
 
 // Run starts the GNparser web service and servies both RESTful API and
 // a website.
-func Run(gnv *gnverify.GNVerify, port int) {
+func Run(gnv gnverify.GNVerify, port int) {
 	var err error
 	e := echo.New()
 
@@ -45,10 +47,24 @@ func Run(gnv *gnverify.GNVerify, port int) {
 	e.Logger.Fatal(e.StartServer(s))
 }
 
-func home(gnv *gnverify.GNVerify) func(echo.Context) error {
+type Data struct {
+	Page     string
+	Input    string
+	Verified []vlib.Verification
+}
+
+func home(gnv gnverify.GNVerify) func(echo.Context) error {
 	return func(c echo.Context) error {
-		return c.Render(http.StatusOK, "layout.html", map[string]interface{}{
-			"name": "Dolly!",
-		})
+		data := Data{Page: "home"}
+		var names []string
+
+		fmt.Printf("data: %#v\n", data)
+		data.Input = c.QueryParam("names")
+		if data.Input != "" {
+			names = strings.Split(data.Input, "\n")
+			fmt.Printf("names: %#v", names)
+			data.Verified = gnv.VerifyBatch(names)
+		}
+		return c.Render(http.StatusOK, "layout", data)
 	}
 }